+++ internal/mock/xml.go
@@ -7,7 +7,10 @@ import (
 	"encoding/xml"
 	"fmt"
 	"io"
+	"strconv"
 	"strings"
+
+	"github.com/issue9/is"
 
 	"github.com/caixw/apidoc/v5/doc"
 	"github.com/caixw/apidoc/v5/internal/locale"
@@ -17,28 +20,20 @@ import (
 type xmlValidator struct {
 	param   *doc.Param
 	decoder *xml.Decoder
-
-	// 按顺序表示的状态
-	// 可以是 [ 表示在数组中，{ 表示在对象，: 表示下一个值必须是属性，空格表示其它状态
-	states []byte
-
-	// 按顺序保存变量名称
-	names []string
+	names   []string // 按顺序保存变量名称
 }
 
 func validXML(p *doc.Request, content []byte) error {
-	if p == nil && bytes.Equal(content, []byte("null")) {
-		return nil
-	}
-
-	if (p.Type == doc.None || p.Type == "") && len(content) == 0 {
-		return nil
+	if len(content) == 0 {
+		if p == nil || p.Type == doc.None {
+			return nil
+		}
+		return message.NewLocaleError("", "", 0, locale.ErrInvalidFormat)
 	}
 
 	validator := &xmlValidator{
 		param:   p.ToParam(),
 		decoder: xml.NewDecoder(bytes.NewReader(content)),
-		states:  []byte{' '}, // 状态有默认值
 		names:   []string{},
 	}
 
@@ -58,40 +53,60 @@ func (validator *xmlValidator) valid() error {
 
 		switch v := token.(type) {
 		case xml.StartElement:
-			// TODO
+			validator.pushName(v.Name.Local)
+			for _, attr := range v.Attr {
+				if err := validator.validValue(attr.Name.Local, attr.Value); err != nil {
+					return err
+				}
+			}
 		case xml.EndElement:
+			validator.popName()
 		case xml.CharData:
-		case xml.Comment:
-		case xml.ProcInst:
-		case xml.Directive:
-		}
-
-		if err != nil {
-			return err
+			if err := validator.validValue("", string(v)); err != nil {
+				return err
+			}
+		case xml.Comment, xml.ProcInst, xml.Directive:
 		}
 	}
 }
 
 // 如果 t == "" 表示不需要验证类型，比如 null 可以赋值给任何类型
-func (validator *xmlValidator) validValue(t doc.Type, v interface{}) error {
-	field := strings.Join(validator.names, ".")
+func (validator *xmlValidator) validValue(k, v string) error {
+	field := strings.Join(validator.names, "/")
 
-	p := validator.find()
+	p := validator.find(k)
 	if p == nil {
+		if k != "" {
+			field += "/" + k
+		}
 		return message.NewLocaleError("", field, 0, locale.ErrNotFound)
 	}
 
-	if t == "" {
-		return nil
+	if p.XMLAttr {
+		field += "/@" + k
 	}
 
-	if p.Type != t {
-		return message.NewLocaleError("", field, 0, locale.ErrInvalidFormat)
+	switch p.Type {
+	case doc.Number:
+		if !is.Number(v) {
+			return message.NewLocaleError("", field, 0, locale.ErrInvalidFormat)
+		}
+	case doc.Bool:
+		if _, err := strconv.ParseBool(v); err != nil {
+			return message.NewLocaleError("", field, 0, locale.ErrInvalidFormat)
+		}
+	case doc.String:
+		return nil
+	case doc.None:
+		if v != "" {
+			return message.NewLocaleError("", field, 0, locale.ErrInvalidValue)
+		}
+	case doc.Object:
 	}
 
 	if p.IsEnum() {
 		for _, enum := range p.Enums {
-			if enum.Value == fmt.Sprint(v) {
+			if enum.Value == v {
 				return nil
 			}
 		}
@@ -99,23 +114,6 @@ func (validator *xmlValidator) validValue(t doc.Type, v interface{}) error {
 	}
 
 	return nil
-}
-
-// 返回当前的状态
-func (validator *xmlValidator) state() byte {
-	return validator.states[len(validator.states)-1]
-}
-
-func (validator *xmlValidator) pushState(state byte) *xmlValidator {
-	validator.states = append(validator.states, state)
-	return validator
-}
-
-func (validator *xmlValidator) popState() *xmlValidator {
-	if len(validator.states) > 0 {
-		validator.states = validator.states[:len(validator.states)-1]
-	}
-	return validator
 }
 
 func (validator *xmlValidator) pushName(name string) *xmlValidator {
@@ -131,9 +129,21 @@ func (validator *xmlValidator) popName() *xmlValidator {
 }
 
 // 如果 names 为空，返回 validator.param
-func (validator *xmlValidator) find() *doc.Param {
+//
+// name 会被附加在 validator.names 之后，进行查询，如果为空，则只查询 validator.names 的值。
+func (validator *xmlValidator) find(name string) *doc.Param {
+	names := make([]string, len(validator.names))
+	copy(names, validator.names)
+	if name != "" {
+		names = append(names, name)
+	}
+
+	if len(names) == 0 || validator.param == nil || names[0] != validator.param.Name {
+		return nil
+	}
+
 	p := validator.param
-	for _, name := range validator.names {
+	for _, name := range names[1:] {
 		found := false
 		for _, pp := range p.Items {
 			if pp.Name == name {
@@ -151,7 +161,145 @@ func (validator *xmlValidator) find() *doc.Param {
 	return p
 }
 
+type xmlBuilder struct {
+	start    xml.StartElement
+	charData string
+	items    []*xmlBuilder
+}
+
+func parseXML(p *doc.Param) (*xmlBuilder, error) {
+	if p == nil || p.Type == doc.None {
+		return nil, nil
+	}
+
+	builder := &xmlBuilder{
+		start: xml.StartElement{
+			Name: xml.Name{
+				Space: p.XMLNSPrefix,
+				Local: p.Name,
+			},
+			Attr: make([]xml.Attr, 0, len(p.Items)),
+		},
+		items: []*xmlBuilder{},
+	}
+
+	if p.Type != doc.Object {
+		switch p.Type {
+		case doc.Bool:
+			builder.charData = fmt.Sprint(generateBool())
+		case doc.Number:
+			builder.charData = fmt.Sprint(generateNumber(p))
+		case doc.String:
+			builder.charData = fmt.Sprint(generateString(p))
+		}
+		return builder, nil
+	}
+
+	for _, item := range p.Items {
+		switch {
+		case item.XMLAttr:
+			v, err := getXMLValue(item)
+			if err != nil {
+				return nil, err
+			}
+
+			builder.start.Attr = append(builder.start.Attr, xml.Attr{
+				Name: xml.Name{
+					Space: item.XMLNSPrefix,
+					Local: item.Name,
+				},
+				Value: fmt.Sprint(v),
+			})
+		case item.XMLExtract:
+			builder.charData = fmt.Sprint(getXMLValue(item))
+		case item.Array:
+			b := builder
+			if item.Wrapped != "" {
+				b = &xmlBuilder{
+					start: xml.StartElement{
+						Name: xml.Name{
+							Space: item.XMLNSPrefix,
+							Local: item.Wrapped,
+						},
+					},
+				}
+			}
+
+			size := generateSliceSize()
+			for i := 0; i < size; i++ {
+				bb, err := parseXML(item)
+				if err != nil {
+					return nil, err
+				}
+				b.items = append(b.items, bb)
+			}
+		default:
+			b, err := parseXML(item)
+			if err != nil {
+				return nil, err
+			}
+			builder.items = append(builder.items, b)
+		}
+	} // end for
+
+	return builder, nil
+}
+
+func (builder *xmlBuilder) encode(e *xml.Encoder) error {
+	if builder == nil {
+		return nil
+	}
+
+	if builder.charData != "" {
+		return e.EncodeElement(builder.charData, builder.start)
+	}
+
+	if err := e.EncodeToken(builder.start); err != nil {
+		return err
+	}
+	for _, item := range builder.items {
+		if err := item.encode(e); err != nil {
+			return err
+		}
+	}
+
+	return e.EncodeToken(xml.EndElement{Name: builder.start.Name})
+}
+
 func buildXML(p *doc.Request) ([]byte, error) {
-	// TODO
-	return nil, nil
+	if p == nil || p.Type == doc.None {
+		return nil, nil
+	}
+
+	builder, err := parseXML(p.ToParam())
+	if err != nil {
+		return nil, err
+	}
+
+	buf := new(bytes.Buffer)
+	e := xml.NewEncoder(buf)
+	if err = builder.encode(e); err != nil {
+		return nil, err
+	}
+
+	if err := e.Flush(); err != nil {
+		return nil, err
+	}
+
+	return buf.Bytes(), nil
+}
+
+func getXMLValue(p *doc.Param) (interface{}, error) {
+	switch p.Type {
+	case doc.None:
+		return "", nil
+	case doc.Bool:
+		return generateBool(), nil
+	case doc.Number:
+		return generateNumber(p), nil
+	case doc.String:
+		return generateString(p), nil
+	default: // doc.Object:
+		return nil, message.NewLocaleError("", "", 0, locale.ErrInvalidFormat)
+	}
 }