+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 
 - 添加 Valid 方法，用于验证文档是否正确；
+- 为 Request 和 Param 添加了五个用于描述 XML 数据的字段；
+- 为 apidoc 元素添加 mimetype，用于指定全局可用的 mimetype 值；
 - 为 Request 和 Param 添加 attr 字段，用于在 content-type 为 xml 的请求中表示属性。
 - 生成 mock 数据的功能；